requirements.txt
psycopg2-binary
aiopg
amqp
asgiref
async-timeout
attrs
autobahn
Automat
beatserver
billiard
certifi
cffi
channels
channels-postgres
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
constantly
croniter
cryptography
daphne
dj-database-url
Django==4.0.5
django-cors-headers
django-filter
django-heroku
djangochannelsrestframework
djangorestframework-simplejwt
django-simple-captcha
djangorestframework
hyperlink
idna
importlib-metadata
incremental
kombu
Markdown
msgpack
msgpack-python
Pillow
prompt-toolkit
psycopg2
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
python-dateutil
python-dotenv
pytz
requests
service-identity
six
sqlparse
txaio
typing-extensions
tzdata
urllib3
vine
wcwidth
whitenoise
zipp
zope.interface
docxtpl
easy-thumbnails
firebase-admin
pyTelegramBotAPI
dialogflow
martor
dialogflow==2.0.0
google-cloud-dialogflow==2.13.0
reportlab
svglib